requirements.txt
# Django framework
Django>=5.0.0,<6.0.0

# Database adapters (uncomment if needed)
# psycopg2-binary>=2.9.9  # For PostgreSQL
# mysqlclient>=2.2.0  # For MySQL

# Web server (for production)
gunicorn>=21.2.0

# Environment variables management
python-dotenv>=1.0.0

# Security
django-cors-headers>=4.3.0

# Form handling
django-crispy-forms>=2.1

# Debugging (development only)
django-debug-toolbar>=4.2.0

# Date/time handling
pytz>=2023.3

# Image processing (if you add this feature later)
# Pillow>=10.0.0

# For handling Excel files (if needed)
# openpyxl>=3.1.2
# xlrd>=2.0.1

# API (for future enhancement)
# djangorestframework>=3.14.0

# AI features
google-generativeai>=0.3.0